Added collision based on Client
* Collision with everything for bullets
* Bug: sometimes collision is triggered when it's not supposed to

+++ src/Map.cpp
@@ -88,7 +88,7 @@ void Map::react() {
 
 void Map::removeMonster(int id) {
 	monsters_.erase(remove_if(monsters_.begin(), monsters_.end(), [&id] (auto& monster) {
-		return monster.getID() == (unsigned int)id;
+		return monster.getID() == id;
 	}));
 }
